refactor: 움직임 상태 추상화
